+++ elements/pt-aboutpage.js
@@ -68,7 +68,8 @@ export default class PTAboutPageElement extends PTPage {
       font-weight: 900;
     }
 
-    :host([data-theme="/themes/fluent-dark.css"]) {
+    :host([data-theme="/themes/fluent-dark.css"]),
+    :host([data-theme="/themes/material-dark.css"]) {
       h1 {
         color: var(--background-color);
       }